Python Code for Space Bugs
I created this game in order to demonstrate the extent that one can use a microbit to create a video game. There is more I could add and different game plays such as using tilt to move the player, but I feel buttons are much easier to control than tilting the entire game.